refactor: apply code review feedback
