+++ tests/unit/Core/Framework/Adapter/Cache/Http/CacheControlListenerTest.php
@@ -5,6 +5,7 @@ namespace Shopware\Tests\Unit\Core\Framework\Adapter\Cache\Http;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
+use Shopware\Administration\Controller\AdministrationController;
 use Shopware\Administration\Framework\Routing\AdministrationRouteScope;
 use Shopware\Core\Framework\Adapter\Cache\Http\CacheControlListener;
 use Shopware\Core\Framework\Adapter\Cache\Http\HttpCacheKeyGenerator;
@@ -172,7 +173,7 @@ class CacheControlListenerTest extends TestCase
     {
         $response = new Response();
         $response->headers->set('cache-control', 'max-age=3600, public, stale-while-revalidate=86400');
-        $response->headers->set('X-Shopware-Cache-Id', 'administration');
+        $response->headers->set(AdministrationController::CACHE_ID_HEADER, AdministrationController::CACHE_ID_ADMINISTRATION);
 
         $request = new Request();
 
@@ -181,7 +182,7 @@ class CacheControlListenerTest extends TestCase
         $subscriber->__invoke(new BeforeSendResponseEvent($request, $response));
 
         static::assertSame('max-age=3600, public, stale-while-revalidate=86400', $response->headers->get('cache-control'));
-        static::assertSame('administration', $response->headers->get('X-Shopware-Cache-Id'));
+        static::assertSame(AdministrationController::CACHE_ID_ADMINISTRATION, $response->headers->get(AdministrationController::CACHE_ID_HEADER));
     }
 
     public function testNonAdministrationHeadersAreModified(): void
